Add system class to replace basis, Modify model
* System class created as a rewrite of the basis class, member functions
will now act on the member variables of the system object, no longer
taking positions etc. as arguments.

* Models are now initialised by passing integrators, systems and
potentials. The keyword approach might have been more user friendly but
this was only applicable to the integrators before, all other parts
were initialised separately. In the interest of making everything
uniform I opted to separate the integrators as well.

* Data handling will still use the data container objects though these
have been modified to inherit the Data class. More work will need to be
done in this area.

* A few of the functions from Potential have been modified as these were
not working as intended.