Update HoppingIntegrator and add Hopping system
All integrators have now been updated a should work. The hopping
integrator is currently only one dimensional and works for only a single
particle at a time. I'm not sure how useful allowing multiple particles
would be, given that interacting particles are currently not available
in any of the methods on here.

Adding 2D capability would be nice but that would require some rewriting
of the tullymodels.py potential. I'm not certain what the best way to
allow both adibatic and diabatic representations of the potential is,
especially in multiple dimensions.

I should mention I have removed the basis files as these have been
wholly replaced by the new systems which in reality are just the same
thing.

Next on the agenda is to add tullymodels 2 and 3 and hope that they
work.