Modify and fix waveintegrators
* SOFT propagator was previously broken, it now works as intended.
* The other wave integrators have been rewritten to be more readable,
though it is possible that 2-dimensional simulations are now broken.
* Following this, visualisations need some work, the 2D wave movie works
very poorly, if at all. Fixing this will allow us to see if the 2D
propagation is working.

* Bonus changes include writing an output file for all simulations and
printing the data fields that have been produced by the integrator.

* Should also note that I decided to temporarily remove the
SOFT_Scattering integrator as I feel it might be possible to have a
general scattering option in the model that can check the end conditions
and determine the scattering result. Perhaps not though, we'll see.