Add SOFT_NonAdiabatic for two energy surfaces
* New SOFT_NonAdiabatic integrator propagates both single and 2 level
systems, could be easily expanded to more by including a numerical
diagonalisation option.
* Slight modifications to Grid1D to work more effectively with the new
integrator.

Should note that I'm not certain it is working as intended yet. It works
for a single surface but might still need some improvements for the
second surface. I'm going to work on a better visualisation framework
for now that will help verify if it is working.