Add 2-level capabilities to wave integrators
* Both SOFT_Propagator and PrimitivePropagator now work for systems with
1 and 2 levels. Both propagators conserve the total energy and the norm
though currently there are some discrepancies between the two. These
include the value of the total energy calculated and the trajectory
followed by the wavepacket, further investigation is required to find
what the issue is.

* The EigenPropagator currently works for only a single energy level but
has been updated to work with the new wave integrator structure, that
is, inherit from the AbstractWavePropagator class.