Add new visualisation framework
Running plot.py with "bokeh serve --show plot.py" now gives a web
interface that allows the user to upload their .end files and
automatically visualise the results.

This is currently in its infancy, only wave1D and traj2D are currently
supported but I think it should be easy to add more in the future.