Add NRPMD and fix Spin Boson model
Nonadiabatic RPMD (NRPMD) as detailed in JCP 150, 244102 (2019)
has been implemented.
The spin boson model potential previously had a error in one the
derivative matrix elements that has now been fixed.
Along with this a couple of modifications were made to the calling of
derivatives in the potential class. This now better matches up with the
standard calling of the potential for multi-level systems.